VideoComponents and CaptureComponents refractored > Rescale bug fix in PacketGenerator
