+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/PatchGenerator/PacketGenerator.java
@@ -48,6 +48,15 @@ public class PacketGenerator {
             prevHashes = new long[tilesX][tilesY];
         }
 
+        if (tilesX >= prevHashes.length || tilesY >= prevHashes[0].length) {
+            // Resize prevHashes array if needed
+            final long[][] newPrevHashes = new long[tilesX][tilesY];
+            for (int i = 0; i < prevHashes.length; i++) {
+                System.arraycopy(prevHashes[i], 0, newPrevHashes[i], 0, prevHashes[i].length);
+            }
+            prevHashes = newPrevHashes;
+        }
+
         final List<CompressedPatch> patches = new ArrayList<>();
 
         for (int ty = 0; ty < tilesY; ty++) {
